+++ NEWS.md
@@ -1,5 +1,7 @@
 # version 0.6-3
 
+* `plot.sf` now emits a warning if `col` does not have length 1 or `nrow(x)`, and requires `pal` to set a palette for factors.
+
 * provide control over legend keys using `key.length` and `key.width`, decrease default key length; #731
 
 * add `as.data.frame` method for `sgbp` objects; #715